♻️ [refactor] #311 : 운동 기록 리스트 컴포넌트 분리
