+++ client/src/pages/RecordPage/index.tsx
@@ -1,9 +1,7 @@
-import React, { useState, useRef, useEffect } from "react";
-import { MdDeleteForever } from "react-icons/md";
-import { AiOutlinePlus } from "react-icons/ai";
+import React, { useRef } from "react";
 import PageTemplate from "@pages/PageTemplate";
 import RoutineScroller from "@components/RoutineScroller";
-import ExerciseRecordItem from "@components/ExerciseRecordItem";
+import ExerciseRecordList from "@components/ExerciseRecordList";
 import RoutineSaveButton from "@components/RoutineSaveButton";
 import exerciseStore from "@stores/exerciseStore";
 import useRecordExercise from "@hooks/query/useRecordExercise";
@@ -13,25 +11,8 @@ import * as s from "./style";
 
 const RecordPage = () => {
   const { recordExercise } = useRecordExercise();
-  const { exerciseList, initExerciseList, createExerciseItem, deleteExerciseItem, fetchRoutine } =
-    exerciseStore();
-
-  const [isIncreaseExercise, setIsIncreaseExercise] = useState<boolean>(false);
+  const { exerciseList, fetchRoutine } = exerciseStore();
   const scrollRef = useRef<HTMLDivElement>(null);
-
-  const handleClickExerciseIncreaseButton = () => {
-    setIsIncreaseExercise(true);
-    createExerciseItem();
-  };
-
-  const handleClickExerciseDecreaseButton = (index: number) => {
-    setIsIncreaseExercise(false);
-    if (exerciseList.length > 1) {
-      deleteExerciseItem(index);
-      return;
-    }
-    initExerciseList();
-  };
 
   const handleClickExerciseSaveButton = () => {
     recordExercise(exerciseList);
@@ -42,42 +23,13 @@ const RecordPage = () => {
     fetchRoutine(routineInfo);
   };
 
-  // 운동 증가 버튼 클릭시 자동 스크롤 다운
-  useEffect(() => {
-    if (!isIncreaseExercise) {
-      return;
-    }
-    setIsIncreaseExercise(false);
-    const targetElement = scrollRef.current;
-    if (targetElement) {
-      targetElement.scrollTop = targetElement.scrollHeight;
-    }
-  }, [isIncreaseExercise]);
-
   return (
     <PageTemplate title="운동" isRoot={false}>
       <s.Wrapper ref={scrollRef}>
         <s.RoutineWrapper>
           <RoutineScroller userId={authStorage.get()} onClickRoutineItem={handleClickRoutineItem} />
         </s.RoutineWrapper>
-        <s.ExerciseListWrapper>
-          {exerciseList.map((_, i) => (
-            // eslint-disable-next-line react/no-array-index-key
-            <s.ExersiceItem key={i}>
-              <s.ExersiceHeader>
-                <p># {i + 1}</p>
-                <s.ExersiceDecreaseButton onClick={() => handleClickExerciseDecreaseButton(i)}>
-                  <MdDeleteForever size={25} />
-                </s.ExersiceDecreaseButton>
-              </s.ExersiceHeader>
-              <ExerciseRecordItem exerciseId={i} />
-            </s.ExersiceItem>
-          ))}
-          <s.ExerciseIncreaseButton onClick={handleClickExerciseIncreaseButton}>
-            <AiOutlinePlus size={20} />
-            <span>운동 추가</span>
-          </s.ExerciseIncreaseButton>
-        </s.ExerciseListWrapper>
+        <ExerciseRecordList scrollRef={scrollRef} />
         <s.SaveButtonWrapper>
           <RoutineSaveButton />
           <s.ExerciseSaveButton onClick={handleClickExerciseSaveButton}>
